chore: rename Comlink parser
* rename Comlink language parser to make room for TypeScript parser integration
