+++ core/comlink/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "comlink_language"
+name = "comlink"
 version = "0.1.0"
 edition = "2021"
 publish = false
