+++ core/comlink/src/comlink_parser/mod.rs
@@ -1,3 +1,5 @@
+//! Parser for the Comlink language.
+
 mod lexer;
 mod syntax;
