+++ core/comlink/src/lib.rs
@@ -1,0 +1,3 @@
+//! This crate provides all tools to work with the Comlink tools.
+
+pub mod comlink_parser;
