--- core/comlink_language/src/lib.rs
@@ -1,3 +1,0 @@
-//! This crate provides all tools to work with the Comlink language.
-
-pub mod parser;
